Split issuer and UserInfo-specific functionality into separate files
